requirements.txt
inky[rpi,example-depends]
Pillow
gpiozero
